requirements.txt
bokeh==2.4.3
graphviz==0.20.1
matplotlib==3.5.2
mysql_connector_repackaged==0.3.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
streamlit==1.11.0
streamlit_echarts==0.4.0
